Move `services` package to `infrastructure` layer
