+++ infrastructure/services/lastfm/lastfm_mock.go
